requirements.txt
requests
expiringdict
python-dateutil
